Simplify the getrescuable function.  Put the actual rescue code in add_partition method.
Its just better to show the possible partitions that can be rescued instead of checking
for the rescuability.  This means that in the list returned by getRescuable, there are
partitions that are not actually rescuable.

Since we dont need to separate the search for rescuable and the commit stuff, we do away
with the rescueable method and put it all in add_partition