Fix traceback if user not root and has not log file defined.
If firstaidkit is run with default config (from configuration.py) the
log file is /var/log/firstaidkit.log.  This could generate a traceback
if the user is not root.  This patch simply make fak fall back on a
file located on the curren directory.  If this further fails print error
and exit with error code.