The -f options implies --nodeps.
When the user uses -f he will be calling only one plugin or a subset
of the total plugins.  This means that the flags will not get set and
the plugin will not run.  This patch fixes this by setting the nodeps
flag.  We also modified the help given to the user.