Add the possibility to pass arguments to plugins.
Given that firstaidkit is fully automated we need a way to modify the
plugin behaviour from the initial command.  This makes it easy for the
plugin developer to get a string with some arguments for his plugin.
It is up to the plugin developer to parse the string.
The arguments will be accessible through self._args variable.