Add the "search for other bootloaders" euristic.
When looking for the place to install GRUB we need to search for
the drives/partitions that dont have other bootloaders in the
first 512 bytes.  This will be done by means of functions that
specifically look for boot loaders.  ATM there is only one
function that will allow an installation allways.  In the futre
we hope to have user defined funcitons for specific bootloaders.