Make most of the subpackages arch independent
Most of the FirstAidKit tool is architecture independent, but the rpm build process doesn't allow noarch package to have arch dependent subpackages. So make the main package a "meta" package which requires noarch subpackages for core engine, arch independent plugins and documentation and also requires proper plugins and parts for the architecture in question.

We will save space by having common stuff created as noarch and we will still be able to exclude some plugins on some architectures.